requirements.txt
PyPDF2==3.0.1
transformers==4.36.0
torch==2.1.0
nltk==3.8.1
python-docx==1.0.1
streamlit==1.29.0
sentence-transformers==2.2.2
protobuf==3.20.0
typing-extensions>=4.5.0 
